requirements.txt
# MILS Assignment 1 - Python Dependencies
# Python 3.8+ recommended

# Core ML Libraries
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0,<2.0
scipy>=1.7.0

# Data Processing
Pillow>=9.0.0
pandas>=1.3.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0

# Training Utilities
tqdm>=4.62.0
scikit-learn>=1.0.0

# Experiment Tracking (optional)
wandb>=0.12.0

# Development Tools
jupyter>=1.0.0
ipython>=7.30.0

# Additional Dependencies for Transformers
einops>=0.4.0  # For Swin Transformer architecture

# CUDA Support (uncomment based on your CUDA version)
# For CUDA 11.6
# torch==1.12.0+cu116 --extra-index-url https://download.pytorch.org/whl/cu116
# torchvision==0.13.0+cu116 --extra-index-url https://download.pytorch.org/whl/cu116

# For CUDA 11.7
# torch==1.13.0+cu117 --extra-index-url https://download.pytorch.org/whl/cu117
# torchvision==0.14.0+cu117 --extra-index-url https://download.pytorch.org/whl/cu117